+++ src/pages/r4-page-channel.js
@@ -1,45 +1,6 @@
-import { html, LitElement } from 'lit'
-import { canEditChannel, readChannel } from '@radio4000/sdk'
+import { html } from 'lit'
 import page from 'page/page.mjs'
-
-// Base class to extend from
-export class ChannelPage extends LitElement {
-	static properties = {
-		channel: { type: Object, state: true },
-		canEdit: { type: Boolean, state: true },
-		// from the router
-		params: { type: Object, state: true },
-		store: { type: Object, state: true },
-		config: { type: Object, state: true },
-	}
-
-	get channelOrigin() {
-		return this.config.singleChannel ? this.config.href : `${this.config.href}/{{slug}}`
-	}
-
-	get tracksOrigin() {
-		if (this.config.singleChannel) {
-			return this.config.href + '/tracks/{{id}}'
-		} else {
-			return this.config.href + '/' + this.params.slug + '/tracks/{{id}}'
-		}
-	}
-
-	connectedCallback() {
-		super.connectedCallback()
-		this.setChannel()
-	}
-
-	// Set channel from the slug in the URL.
-	async setChannel() {
-		this.channel = (await readChannel(this.params.slug)).data
-		this.canEdit = await canEditChannel(this.params.slug)
-	}
-
-	render() {
-		return html``
-	}
-}
+import { ChannelPage } from './base-channel'
 
 export default class R4PageChannel extends ChannelPage {
 	render() {
@@ -111,11 +72,11 @@ export default class R4PageChannel extends ChannelPage {
 		}
 	}
 
-	onDialogClose({target}) {
+	onDialogClose({ target }) {
 		const name = target.getAttribute('name')
 		if (name === 'track') {
 			if (this.config.singleChannel) {
-				page('/tracks')
+				page('/track')
 			} else {
 				page(`/${this.params.slug}/tracks`)
 			}
